+++ stm/usrsw.h
@@ -1,0 +1,7 @@
+#ifndef __USRSW_H__
+#define __USRSW_H__
+void sw_init(void);
+int sw_get(void);
+
+mp_obj_t pyb_sw(void);
+#endif //__USRSW_H__
